Added and changed html files
I added files public document and shared document and changed the name for document to private document to allow users different actions based on where they are looking at a file. For public document they are able to view only. On private they have access to view, edit, delete and set to public or private if its already public. For a shared document they are able to edit and view. I also made the title larger on each of the main pages and made the buttons look a little better.